- add a new model type for nostr, refactor the code for nostr chat - implement the toast error for empty nostr account
