+++ apps/mobile/store/accountBuilder.ts
@@ -3,12 +3,8 @@ import uuid from 'react-native-uuid'
 import { create } from 'zustand'
 
 import { type EntropyType } from '@/types/logic/entropy'
-import {
-  type Account,
-  type DM,
-  type Key,
-  type Secret
-} from '@/types/models/Account'
+import { type Account, type Key, type Secret } from '@/types/models/Account'
+import type { NostrDM } from '@/types/models/Nostr'
 
 type AccountBuilderState = {
   name: Account['name']
@@ -247,7 +243,7 @@ const useAccountBuilderStore = create<
         deviceNpub: '',
         deviceNsec: '',
         trustedMemberDevices: [],
-        dms: [] as DM[],
+        dms: [] as NostrDM[],
         lastUpdated: new Date(),
         syncStart: new Date()
       }
