+++ apps/mobile/types/models/Account.ts
@@ -2,6 +2,7 @@ import { type Network } from '../settings/blockchain'
 import { type Address } from './Address'
 import { type Transaction } from './Transaction'
 import { type Utxo } from './Utxo'
+import type { NostrAccount } from '@/types/models/Nostr'
 
 export type PolicyType = 'singlesig' | 'multisig' | 'watchonly'
 
@@ -51,20 +52,6 @@ export type Key = {
   derivationPath?: string
 }
 
-export type DM = {
-  id: string
-  author: string
-  created_at: number
-  description: string
-  event: string
-  label: number
-  content: {
-    description: string
-    created_at: number
-    pubkey?: string
-  }
-}
-
 export type Account = {
   id: string
   name: string
@@ -91,40 +78,5 @@ export type Account = {
   lastSyncedAt?: Date
   syncStatus: SyncStatus
   syncProgress?: SyncProgress
-  nostr: {
-    commonNpub: string
-    commonNsec: string
-    relays: string[]
-    autoSync: boolean
-    lastBackupFingerprint?: string
-    deviceNpub?: string
-    deviceNsec?: string
-    trustedMemberDevices: string[]
-    dms: DM[]
-    lastUpdated: Date
-    syncStart: Date
-    relayStatuses?: Record<string, 'connected' | 'connecting' | 'disconnected'>
-  }
-}
-
-export type NostrAccount = {
-  commonNpub: string
-  commonNsec: string
-  relays: string[]
-  autoSync: boolean
-  lastBackupFingerprint?: string
-  deviceNpub?: string
-  deviceNsec?: string
-  dms?: DM[]
-  members?: NostrMember[]
-  syncStart: Date
-  lastProtocolEOSE?: number
-  lastDataExchangeEOSE?: number
-  lastUpdated: Date
-  trustedMemberDevices: string[]
-}
-
-export type NostrMember = {
-  npub: string
-  color?: string
+  nostr: NostrAccount
 }
